requirements.txt
numpy~=1.19.3
pandas~=1.4.1
scikit-image~=0.18.3
opencv
pillow~=8.2.0
pytesseract~=0.3.9
dash~=2.3.0
opencv-python~=4.5.5.64
paddleocr~=2.4.0.3
matplotlib~=3.4.2
scipy~=1.7.1
flask~=1.1.2